+++ packages/associative/src/index.ts
@@ -17,5 +17,6 @@ export * from "./rename-keys";
 export * from "./select-keys";
 export * from "./sorted-map";
 export * from "./sorted-set";
+export * from "./sparse-set";
 export * from "./union";
 export * from "./without-keys";