+++ packages/geom-isec/src/ray-line.ts
@@ -13,7 +13,7 @@ export const intersectRayLine = (
     const bay = b[1] - a[1];
     const d = dir[0] * bay - dir[1] * bax;
     if (eqDeltaFixed(d, 0)) {
-        return { type: IntersectionType.NONE };
+        return NONE;
     }
     const arx = a[0] - rpos[0];
     const ary = a[1] - rpos[1];