+++ packages/geom-isec/test/index.ts
@@ -1,6 +1,0 @@
-// import * as assert from "assert";
-// import * as gi from "../src/index";
-
-describe("geom-isec", () => {
-    it("tests pending");
-});