+++ packages/geom/src/api.ts
@@ -437,6 +437,19 @@ export class Quad extends APC implements IHiccupShape {
         return ["polygon", this.attribs, this.points];
     }
 }
+export class Quad3 extends APC implements IHiccupShape {
+    get type() {
+        return Type.QUAD3;
+    }
+
+    copy() {
+        return new Quad3(copyVectors(this.points), { ...this.attribs });
+    }
+
+    toHiccup() {
+        return ["polygon", this.attribs, this.points];
+    }
+}
 
 export class Quadratic extends APC implements IHiccupShape, IHiccupPathSegment {
     get type() {