+++ packages/malloc/src/api.ts
@@ -1,28 +1,4 @@
-import { IRelease, TypedArray } from "@thi.ng/api";
-
-export const enum Type {
-    U8,
-    U8C,
-    I8,
-    U16,
-    I16,
-    U32,
-    I32,
-    F32,
-    F64
-}
-
-export const SIZEOF = {
-    [Type.U8]: 1,
-    [Type.U8C]: 1,
-    [Type.I8]: 1,
-    [Type.U16]: 2,
-    [Type.I16]: 2,
-    [Type.U32]: 4,
-    [Type.I32]: 4,
-    [Type.F32]: 4,
-    [Type.F64]: 8
-};
+import { IRelease, Type, TypedArray } from "@thi.ng/api";
 
 export interface MemBlock {
     addr: number;