+++ packages/iterators/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.19](https://github.com/thi-ng/umbrella/compare/@thi.ng/iterators@5.0.18...@thi.ng/iterators@5.0.19) (2019-05-22)
+
+**Note:** Version bump only for package @thi.ng/iterators
+
+
+
+
+
 ## [5.0.18](https://github.com/thi-ng/umbrella/compare/@thi.ng/iterators@5.0.17...@thi.ng/iterators@5.0.18) (2019-04-26)
 
 **Note:** Version bump only for package @thi.ng/iterators