+++ packages/pointfree/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/thi-ng/umbrella/compare/@thi.ng/pointfree@1.0.13...@thi.ng/pointfree@1.0.14) (2019-05-22)
+
+
+### Bug Fixes
+
+* **pointfree:** update safeMode handling ([d27bcba](https://github.com/thi-ng/umbrella/commit/d27bcba))
+
+
+
+
+
 ## [1.0.13](https://github.com/thi-ng/umbrella/compare/@thi.ng/pointfree@1.0.12...@thi.ng/pointfree@1.0.13) (2019-04-26)
 
 **Note:** Version bump only for package @thi.ng/pointfree