+++ packages/rstream-dot/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.26](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-dot@1.0.25...@thi.ng/rstream-dot@1.0.26) (2019-05-22)
+
+**Note:** Version bump only for package @thi.ng/rstream-dot
+
+
+
+
+
 ## [1.0.25](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-dot@1.0.24...@thi.ng/rstream-dot@1.0.25) (2019-04-26)
 
 **Note:** Version bump only for package @thi.ng/rstream-dot