+++ packages/transducers-binary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.12](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-binary@0.3.11...@thi.ng/transducers-binary@0.3.12) (2019-05-22)
+
+**Note:** Version bump only for package @thi.ng/transducers-binary
+
+
+
+
+
 ## [0.3.11](https://github.com/thi-ng/umbrella/compare/@thi.ng/transducers-binary@0.3.10...@thi.ng/transducers-binary@0.3.11) (2019-04-26)
 
 **Note:** Version bump only for package @thi.ng/transducers-binary